pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.18.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.2",
    "python-dotenv>=1.1.0",
]
